app/requirements.txt
Flask==1.0.2
google-cloud==0.34.0
google-cloud-storage==1.13.2
